crud/requirements.txt
django
django-rest-framework
django-filter
psycopg2